backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
fastapi-users[sqlalchemy]==12.1.2
fastapi-users[oauth]==12.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis for caching and real-time
redis==5.0.1
aioredis==2.0.1

# WebSocket support
websockets==12.0

# File handling and AWS
boto3==1.34.0
pillow==10.1.0
python-magic==0.4.27

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Vector database for future RAG
pgvector==0.2.4

# AI/ML (for future phases)
openai==1.3.7
numpy==1.25.2